+++ lib/main.dart
@@ -30,22 +30,22 @@ import 'package:timecop/l10n.dart';
 import 'package:timecop/screens/dashboard/DashboardScreen.dart';
 import 'package:timecop/themes.dart';
 
-import 'package:timecop/data_providers/database_provider.dart';
+/*import 'package:timecop/data_providers/database_provider.dart';
 import 'package:timecop/data_providers/shared_prefs_settings_provider.dart';
 Future<void> main() async {
   WidgetsFlutterBinding.ensureInitialized();
   final SettingsProvider settings = await SharedPrefsSettingsProvider.load();
   final DataProvider data = await DatabaseProvider.open();
   await runMain(settings, data);
-}
+}*/
 
-/*import 'data_providers/mock_data_provider.dart';
+import 'data_providers/mock_data_provider.dart';
 import 'data_providers/mock_settings_provider.dart';
 Future<void> main() async {
   final SettingsProvider settings = MockSettingsProvider();
   final DataProvider data = MockDataProvider(Locale.fromSubtags(languageCode: "en"));
   await runMain(settings, data);
-}*/
+}
 
 Future<void> runMain(SettingsProvider settings, DataProvider data) async {
   // setup intl date formats?