+++ lib/screens/timer/TimerEditor.dart
@@ -26,6 +26,7 @@ import 'package:timecop/components/ProjectColour.dart';
 import 'package:timecop/l10n.dart';
 import 'package:timecop/models/project.dart';
 import 'package:timecop/models/timer_entry.dart';
+import 'package:timecop/models/clone_time.dart';
 
 class TimerEditor extends StatefulWidget {
   final TimerEntry timer;
@@ -39,8 +40,13 @@ class TimerEditor extends StatefulWidget {
 
 class _TimerEditorState extends State<TimerEditor> {
   TextEditingController _descriptionController;
+
   DateTime _startTime;
   DateTime _endTime;
+
+  DateTime _oldStartTime;
+  DateTime _oldEndTime;
+
   Project _project;
   FocusNode _descriptionFocus;
   final _formKey = GlobalKey<FormState>();
@@ -70,12 +76,16 @@ class _TimerEditorState extends State<TimerEditor> {
     super.dispose();
   }
 
-  @override
   void setStartTime(DateTime dt) {
-    setState(() { _startTime = dt; });
-    if(_startTime.isAfter(_endTime)){
-      setState(() { _endTime = _startTime; });
-    }
+    assert(dt != null);
+    setState(() {
+      // adjust the end time to keep a constant duration if we would somehow make the time negative
+      if(_oldEndTime != null && dt.isAfter(_oldStartTime)) {
+        Duration d = _oldEndTime.difference(_oldStartTime);
+        _endTime = dt.add(d);
+      }
+      _startTime = dt;
+    });
   }
 
   @override
@@ -87,9 +97,6 @@ class _TimerEditorState extends State<TimerEditor> {
       body: Form(
         key: _formKey,
         child: ListView(
-          //mainAxisAlignment: MainAxisAlignment.start,
-          //mainAxisSize: MainAxisSize.max,
-          //crossAxisAlignment: CrossAxisAlignment.stretch,
           children: <Widget>[
             BlocBuilder<ProjectsBloc, ProjectsState>(
               builder: (BuildContext context, ProjectsState projectsState) => Padding(
@@ -147,10 +154,13 @@ class _TimerEditorState extends State<TimerEditor> {
               title: Text(L10N.of(context).tr.startTime),
               trailing: Text(_dateFormat.format(_startTime)),
               onTap: () async {
-                await DatePicker.showDateTimePicker(
+                _oldStartTime = _startTime.clone();
+                _oldEndTime = _endTime.clone();
+                DateTime newStartTime = await DatePicker.showDateTimePicker(
                   context,
                   currentTime: _startTime,
-//                  onChanged: (DateTime dt) => setStartTime(dt),
+                  maxTime: _endTime == null ? DateTime.now() : null,
+                  onChanged: (DateTime dt) => setStartTime(dt),
                   onConfirm: (DateTime dt) => setStartTime(dt),
                   theme: DatePickerTheme(
                     cancelStyle: Theme.of(context).textTheme.button,
@@ -159,6 +169,14 @@ class _TimerEditorState extends State<TimerEditor> {
                     backgroundColor: Theme.of(context).scaffoldBackgroundColor,
                   )
                 );
+
+                // if the user cancelled, this should be null
+                if(newStartTime == null) {
+                  setState(() {
+                    _startTime = _oldStartTime;
+                    _endTime = _oldEndTime;
+                  });
+                }
               },
             ),
             Slidable(
@@ -168,11 +186,12 @@ class _TimerEditorState extends State<TimerEditor> {
                 title: Text(L10N.of(context).tr.endTime),
                 trailing: Text(_endTime == null ? "—" : _dateFormat.format(_endTime)),
                 onTap: () async {
-                  await DatePicker.showDateTimePicker(
+                _oldEndTime = _endTime.clone();
+                DateTime newEndTime = await DatePicker.showDateTimePicker(
                     context,
                     currentTime: _endTime,
                     minTime: _startTime,
-//                    onChanged: (DateTime dt) => setState(() => _endTime = dt),
+                    onChanged: (DateTime dt) => setState(() => _endTime = dt),
                     onConfirm: (DateTime dt) => setState(() => _endTime = dt),
                     theme: DatePickerTheme(
                       cancelStyle: Theme.of(context).textTheme.button,
@@ -181,6 +200,13 @@ class _TimerEditorState extends State<TimerEditor> {
                       backgroundColor: Theme.of(context).scaffoldBackgroundColor,
                     )
                   );
+
+                // if the user cancelled, this should be null
+                if(newEndTime == null) {
+                  setState(() {
+                    _endTime = _oldEndTime;
+                  });
+                }
                 },
               ),
               secondaryActions: